[Refactor] #1644 Separated use-execution.cpp/h from cmd-usestaff.cpp/h
