+++ src/cmd-item/cmd-item.cpp
@@ -18,7 +18,6 @@
 #include "cmd-item/cmd-eat.h"
 #include "cmd-item/cmd-quaff.h"
 #include "cmd-item/cmd-read.h"
-#include "cmd-item/cmd-usestaff.h"
 #include "cmd-item/cmd-zaprod.h"
 #include "cmd-item/cmd-zapwand.h"
 #include "combat/shoot.h"
@@ -39,6 +38,7 @@
 #include "object-hook/hook-magic.h"
 #include "object-use/quaff-execution.h"
 #include "object-use/read-execution.h"
+#include "object-use/use-execution.h"
 #include "object/item-tester-hooker.h"
 #include "object/item-use-flags.h"
 #include "perception/identification.h"
