+++ src/cmd-item/cmd-usestaff.h
@@ -4,5 +4,4 @@
 
 struct player_type;
 int staff_effect(player_type *player_ptr, OBJECT_SUBTYPE_VALUE sval, bool *use_charge, bool powerful, bool magic, bool known);
-void exe_use_staff(player_type *player_ptr, INVENTORY_IDX item);
 void do_cmd_use_staff(player_type *player_ptr);
