+++ src/object-use/use-execution.h
@@ -1,0 +1,6 @@
+﻿#pragma once
+
+#include "system/angband.h"
+
+struct player_type;
+void exe_use_staff(player_type *player_ptr, INVENTORY_IDX item);
